Hydra API change in GetNumElements to return a size_t instead of int
(Internal change: 1871209)

+++ pxr/imaging/lib/hd/bufferArrayRange.h
@@ -87,7 +87,7 @@ public:
     virtual int GetIndex() const = 0;
 
     /// Returns the number of elements
-    virtual int GetNumElements() const = 0;
+    virtual size_t GetNumElements() const = 0;
 
     /// Returns the version of the buffer array.
     virtual size_t GetVersion() const = 0;
